+++ tests/functional/test_functional.py
@@ -58,7 +58,7 @@ class TestFunctional(unittest.TestCase):
     def test_write_resources_in_region(self):
         """It is sufficient for this not to throw an exception."""
         self.wanderer.write_resources(
-            regions=["us-east-1"], exclude_resources=["ec2:image", "ec2:snapshot", "iam:policy"]
+            regions=["eu-west-1"], exclude_resources=["ec2:image", "ec2:snapshot", "iam:policy"]
         )
 
     def test_write_resource_type(self):